Refactor volleyball game management and update CreateGame component: - Updated ManageCalendar to change game period status from '1º Tempo' to 'in_progress'. - Removed CreateGame component to streamline game creation process. - Enhanced VolleyballScore component with improved game detail fetching and state management. - Added logic to update game status to 'live' and period to 'in_progress' upon scoring. - Improved VolleyballHighlights and VolleyballPoints components to handle set scoring and game updates more effectively. - Updated types for VolleyballGame to include new status management. These changes enhance the overall functionality and user experience in managing volleyball games.
